Add files via upload
The scraped links are utilized in these python scripts to go to each job postings web page and to copy the html of the page to a new column. this is repeated until there is no more job links to go to. The process is automated so you can just watch it go through everything. It looks cool but itll take about an hour to 2 hours for 3000 postings.